pyproject.toml
[build-system]
requires = ["setuptools>=45.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "model-cli-autocomplete"
version = "0.1.0"
description = "AI-powered CLI autocomplete tool with Ollama integration"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "argcomplete>=2.0.0",
    "python-dotenv>=0.19.0",
    "prompt-toolkit>=3.0.0",
]

[project.optional-dependencies]
training = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "accelerate>=0.20.0",
    "datasets>=2.12.0",
    "peft>=0.4.0",
    "axolotl>=0.3.0",
]